StoryBoard Deferred Processors
This patch adds a puppet module that will start up N worker threads
to handle messages sent to the deferred processing queue. It does so
by creating a "parent" service called storyboard-workers, which
is treated as a regular upstart service. In a pre-start script,
it also starts up N "child" services called storyboard-worker, which
listen for the DOWN state on the parent service.

Upstart start-on and stop-on directives are autoconfigured based on
whether this particular service is running storyboard::mysql or
storyboard::rabbit.

Depends on 112407

Change-Id: I5565cbf8062457d343d3e02dbfaae2852a359d91
Story: 96