Disabled rabbit repo management.
Turns out the rabbit module manages its own APT repository, which
can end up with some unexpected results on trusty. This disables
that and defaults back to using the trusty upstream.

After this patch lands, we're going to have to manually update
storyboard.o.o to remove the ppa and reinstall rabbit.

Change-Id: Ic5ada12e730845e550d1beb934c536955b77ef16