Rabbit repository management
If we are running on anything before trusty, we need to manually
manage the repository because the available version of rabbitmq
is not recent enough for our module. For trusty and later, the
available version is too recent for us to use, so we have to turn
off repo management.

This is volatile, and needs a more permanent solution.

Change-Id: I0db87d6c5cbecee8575d91e6f88ef43a947967ff