Apache Config is now an optional module.
This patch moves the vhost configuration for storyboard into two
separate modules that may be optionally included.

DO NOT MERGE. There is an outstanding question on how to trigger
an apache restart when content in storyboard::application changes.
Since we cannot assume that apache is installed, we cannot notify
service['httpd'], nor can we modify this service to subscribe
to changes in our application and/or virtual host configuration.
Suggestions welcome.

Change-Id: If237fbffa644fafa1788987a6eb7a82b5207b03b