StoryBoard is now installed into a Virtualenv
In order to isolate storyboard's code from other parts of the host system,
it is now installed into a virtual environment and run from there. The
virtual environment directory is tagged with the version of python, for
future clean upgrades.

Note: After this update, an admin should probably go remove the old
install directory on the server.

Change-Id: Ia93cf09c35189c22a208a0e917ef7ee5cbd4b510