Clean the webclient unpack dir safely.
Don't rely on a puppet file resource to purge the unpack dir as this
will trigger on every puppet run then copy the empty unpack dir to the
deployment dir. Instead we call rm prior to tar during the unpack exec
so that we only ever clean the dir if we are unpacking.

Change-Id: I80b4faf54220d88e610e2a224dea47af4133e225