Remove unused puppi dependency
We assert a dependency on this puppi puppet module but then never use
it. Looking at the git log we used it at one time to download tarballs
but then stopped as it wasn't reliable enough.

Clean this up to make puppet 4 happier about this module not having
puppi available in production.

Change-Id: Ifed308229b92035e8b62d8b3677ca698639c6804